+++ parser/src/main/kotlin/parser/ExpressionParser.kt
@@ -1,24 +1,20 @@
 package parser
 
 import ast.AST
-import ast.DivisionNode
-import ast.ExpressionNode
-import ast.IdentifierNode
-import ast.LiteralNode
-import ast.OperatorNode
-import ast.OperatorType
-import ast.ProductNode
-import ast.SubtractionNode
-import ast.SumNode
+import parser.factory.OperationStrategyFactory
+import parser.factory.TokenHandlerFactory
 import parser.result.ParserResult
 import parser.result.SuccessResult
+import parser.utils.OperatorProvider
 import parser.utils.at
 import parser.utils.isOutOfBounds
 import parser.utils.nextIndex
 import token.Token
-import token.TokenType
 
-class ExpressionParser : Parser {
+class ExpressionParserV3(
+    private val strategyFactory: OperationStrategyFactory,
+    private val tokenHandlerFactory: TokenHandlerFactory,
+) : Parser {
     override fun parse(
         tokens: List<Token>,
         currentIndex: Int,
@@ -36,7 +32,12 @@ class ExpressionParser : Parser {
         var lastIndex = leftResult.lastValidatedIndex
         var currentOperatorIndex = nextIndex(lastIndex)
 
-        while (!isOutOfBounds(tokens, currentOperatorIndex) && isPrimaryOperator(tokens, currentOperatorIndex)) {
+        while (!isOutOfBounds(tokens, currentOperatorIndex) &&
+            OperatorProvider.isPrimaryOperator(
+                tokens,
+                currentOperatorIndex,
+            )
+        ) {
             val operator = at(tokens, currentOperatorIndex)
             val rightResult = parseTerm(tokens, nextIndex(currentOperatorIndex))
 
@@ -61,7 +62,12 @@ class ExpressionParser : Parser {
 
         var currentOperatorIndex = nextIndex(lastIndex)
 
-        while (!isOutOfBounds(tokens, currentOperatorIndex) && isTermOperator(tokens, currentOperatorIndex)) {
+        while (!isOutOfBounds(tokens, currentOperatorIndex) &&
+            OperatorProvider.isTermOperator(
+                tokens,
+                currentOperatorIndex,
+            )
+        ) {
             val operator = at(tokens, currentOperatorIndex)
             val rightResult = parserFactor(tokens, nextIndex(currentOperatorIndex))
 
@@ -81,84 +87,9 @@ class ExpressionParser : Parser {
         currentIndex: Int,
     ): SuccessResult {
         val currentToken = at(tokens, currentIndex)
-
-        return when (currentToken.type) {
-            TokenType.IDENTIFIER -> {
-                SuccessResult(
-                    IdentifierNode(
-                        currentToken.value,
-                        start = currentToken.start,
-                        end = currentToken.end,
-                    ),
-                    currentIndex,
-                )
-            }
-
-            TokenType.NUMBER -> {
-                SuccessResult(
-                    LiteralNode(
-                        currentToken.value.toDouble(),
-                        start = currentToken.start,
-                        end = currentToken.end,
-                    ),
-                    currentIndex,
-                )
-            }
-
-            TokenType.STRING -> {
-                SuccessResult(
-                    LiteralNode(
-                        currentToken.value,
-                        start = currentToken.start,
-                        end = currentToken.end,
-                    ),
-                    currentIndex,
-                )
-            }
-
-            TokenType.OPENPARENTHESIS -> {
-                val expressionResult = parsePrimaryExpression(tokens, nextIndex(currentIndex))
-                val expression = expressionResult.value
-
-                val rightParen = at(tokens, nextIndex(expressionResult.lastValidatedIndex))
-                if (rightParen.type != TokenType.CLOSEPARENTHESIS) {
-                    throw Exception("Expected a right parenthesis")
-                }
-
-                SuccessResult(
-                    expression,
-                    nextIndex(expressionResult.lastValidatedIndex),
-                )
-            }
-
-            else -> throw Exception("Unexpected token: ${currentToken.value}")
-        }
-    }
-
-    private fun isPrimaryOperator(
-        tokens: List<Token>,
-        operatorIndex: Int,
-    ): Boolean {
-        return when (at(tokens, operatorIndex).type) {
-            TokenType.SUM,
-            TokenType.SUBTRACTION,
-            -> true
-
-            else -> false
-        }
-    }
-
-    private fun isTermOperator(
-        tokens: List<Token>,
-        operatorIndex: Int,
-    ): Boolean {
-        return when (at(tokens, operatorIndex).type) {
-            TokenType.MULTIPLICATION,
-            TokenType.DIVISION,
-            -> true
-
-            else -> false
-        }
+        val handler = tokenHandlerFactory.getHandler(currentToken.type, this)
+        val astNode = handler.handleToken(tokens, currentIndex)
+        return SuccessResult(astNode, currentIndex)
     }
 
     private fun buildOperationNode(
@@ -166,36 +97,7 @@ class ExpressionParser : Parser {
         right: AST,
         operator: Token,
     ): AST {
-        when (operator.type) {
-            TokenType.SUM -> return SumNode(
-                left as ExpressionNode,
-                right as ExpressionNode,
-                OperatorNode(operator.start, operator.end, OperatorType.SUM),
-                left.getStart(),
-                right.getEnd(),
-            )
-            TokenType.SUBTRACTION -> return SubtractionNode(
-                left as ExpressionNode,
-                right as ExpressionNode,
-                OperatorNode(operator.start, operator.end, OperatorType.SUB),
-                left.getStart(),
-                right.getEnd(),
-            )
-            TokenType.MULTIPLICATION -> return ProductNode(
-                left as ExpressionNode,
-                right as ExpressionNode,
-                OperatorNode(operator.start, operator.end, OperatorType.MUL),
-                left.getStart(),
-                right.getEnd(),
-            )
-            TokenType.DIVISION -> return DivisionNode(
-                left as ExpressionNode,
-                right as ExpressionNode,
-                OperatorNode(operator.start, operator.end, OperatorType.DIV),
-                left.getStart(),
-                right.getEnd(),
-            )
-            else -> throw Exception("Unexpected operator: ${operator.value}")
-        }
+        val strategy = strategyFactory.getStrategy(operator.type)
+        return strategy.execute(left, right, operator)
     }
 }
